Moved SDF prototype files to their own folder to make it easier to add more prototypes.
