+++ src/lib.rs
@@ -2,3 +2,5 @@ pub mod minivector;
 pub mod sdf;
 pub mod serialization;
 pub mod sparse_sdf;
+pub mod vulkan_base;
+pub mod vulkan_helpers;
